Move image build/push from Makefile into Jenkinsfile release
The make targets for building and pushing the image are currently no longer
needed. This commit removes the obsolete code. This commit also re-arranges
the sequence of image building/pushing within the Jenkinsfile for better
fail-safety. If an error occurs during pushing the github releases stay
untouched and do not need to be rolled back.