Extend release shell scripts with extension points
This commit adds a way to extend the release shell scripts with
so-called custom release scripts. The idea is to source a file which
defines methods that might be called by the release scripts. The
release scripts check if such a method exists. If so, the function
will be called during the release.

Since there might be more than one method might be necessary to
complete a call, each release method may define multiple extension
points methods.

Custom release scripts should then implement these extensions points
and to the special work of which the original release scripts
weren't designed to (f. i. editing special YAML files which only may
exist in one repository)