#40 Add doguRegistrator method for registering new dogu versions, extract interfaces
The current implementation of registering a dogu differs from the cesapp
implementation which registers also existing dogus and a new version. The key
difference is that the dogu operator implementation checks if the dogu is
installed. In the case of the previous implementation that would lead to an
early exit.

This commit adds another method that will be used instead. Also an error will be
returned if the dogu seems not to be installed in the first place (in terms of
a missing registry key).

Also this commit extracts the broad load of interfaces in the controllers
package into its own file.