#40 Remove Applier scheme, add DoguUpgradeManager tests
The scheme, which the applier-lib generates and returns, does not seem to be
needed for the Dogu Operator AT ALL. The scheme is only needed when applying
new CRDs (namely, the Dogu CRD) which is the subject of k8s-ces-setup and not
that of this Dogu Operator.

This commit removes the calls and error handling in question.