#40 Fix requeue behaviour after non-requeueable errors, fix upgrade state change
This commit fixes the previous requeue behaviour which was buggy when
unqueueable errors occurred. Upon every error the original error was returned
which subsequently lead to another reconciliation loop. The desired behaviour is
instead to NOT requeue the resource change because the error already indicates
as non-requeueable. Instead the administrator needs to inspect the cluster state
in order to fix the dogu upgrade.

For the operations Install and Delete this is no big deal because either way all
previous resourced can be deleted without losing any production data because
- either there is no data yet
- or all data are supposed to be deleted

This commit also moves the upgrade's dogu state change to an earlier stage in
order to support the actual notion of errors that may or may not be requeueable.
The previous behavior f. i. did not change the state when a premises check
failed and the dogu state kept stuck on "installed" when it was supposed to be
set to "upgrading"