#40 Switch deployment strategy to recreate
The current build process acts sometimes weird and has two crashloop pods for a
deployment when there should only be one. The reason is that the default
deployment strategy goes for rolling updates
which does not bode well during the startup where
one will crash.

This commit changes the deployment strategy to recreate the pod which has
the drawback of downtime. This drawback is only a minor hindrance because the
operator has a fast spinup time.