#47 Add exec pod factory
ExecPods need to be initialized during the run-time of the operator, i. e. with
a dogu resource at hands. Under that circumstance it is not possible to create
an execPod during the manager creation time. In order to circumvent any
complications during testing a factory will help out here -- which is testable
and mockable as well.