#62 set correct object key during exposed service creation
The object key of the dogu resource was used as-is which is plain wrong. When
the operator tries to get the exposed service resource the resource will not be
returned because it is not found.

The resource generator creates quite specific object keys which includes the
port. This commit changes to object to the correct generated key in order to
find the exposed service later-on.