#66 Execute pre-upgrade script with pipes on copy-fail
This commit adds a fallback strategy for executing pre-upgrade scripts to solve
the problem of file system permission mongering when the script is owned by root
and is executable by the world.

During a dogu upgrade the old dogu's container filesystem layout looks like
this after the exec pod copied

```
ls -lha /
drwxr-xr-x    1 root     root        4.0K Dec 13 10:47 .
-rwxrwxr-x    1 root     root         704 Oct 12 14:25 pre-upgrade.sh
...

ls -lha /tmp/dogu-reserved/
drwxrwsr-x    3 root     doguuser    1.0K Dec 13 10:48 .
-rwxr-xr-x    1 doguuser doguuser     704 Dec 13 10:48 pre-upgrade.sh
...
```

Because of how the permission are set the newer `pre-upgrade.sh` cannot be
copied because of two reasons:
1. the file exists and belongs to root (cannot be overwritten if it exists)
2. the directory / belongs to root (current dogu user cannot create a file at
   all)

This commit adds a second strategy that hints which exec implementation should
run. The solution heavily relies on the proper detection of the copy error. If
detected the whole script will instead be piped against the script's desired
script interpreter.

This implementation is less complicated to test and implement compared
to a full strategy pattern implementation in their own structs which complicates
testing even further.