Fix nil pointer dereference for dogus without Startup probe
Dogus that do not define a health check of type `state` in their dogu.json do
not receive a startup probe. Doing upgrades on these dogus lead to a nil error
because the startup probe is of a pointer type and will not be set during the
assignment.

This commit adds nil pointer checks to avoid setting probes where those cannot
be set.