#66 Revert complex pre-upgrade logic for simple script execution
The previous commits introduced a complex business logic which did not take into
account that dogu developers may create pre-upgrade scripts which reside in any
sub-directories. In cases, where the dogu file system layout does not allow for
file copying or even directory creation, the execution of pre-upgrade scripts is
prone to fail by design.

This commit introduces a set of conventions for pre-upgrade scripts as
documented in the docs/ directory.