Boyscouting: remove superfluous dogu.name label from dogu example resources
The label dogu.name is only used to uniquely identify dogu pods. That label is
necessary to differ from two different pods during the upgrade of the same dogu.
Thus this label is unnecessary for dogu resources because those can be addressed
simply by the resource `dogu` and its name.

Co-authored-by: Niklas <niklas.hinze@cloudogu.com>, Alex <alexander.dammeier@cloudogu.com>