#74 Add chown init container for dogus with volumes.
Some dogus like teamscale define volumes with uid and gid but does not chown the directories in the dockerfile. If the container runs with root the volumes are owned with the ids 0. This can prevent dogus (e.g. teamscale) to execute the application with the application user (from the startup.sh).