#83 Dogu registration refactoring.
Enable the dogu only in the last step of the registration, so it won't be skipped in a future registration process if an error occurs (e.g. key generation). Furthermore, add functionality to handle existing private keys to stabilize the process. In this situation only a new public key will be generated.