#118 configure chown init container image
Previously, additionally required images are mentioned in the dogu operators
k8s resource file set as annotation as well as hardcoded within the code.

This commit removes hardcoded references in favor of a Configmap that consists
of a constant key and the default image reference. This image reference can be
replaced by the user before applying the dogu operator for better customize-
ability.

The image validation regexp with courtesy of https://stackoverflow.com/a/39672069/12529534