#248 AdditionalMounts: Use user and group id from dogu
Without it, the local config gets the owner id from the init container on first start.
For most dogus, this is okay because nearly all ids are 1000 and do not conflict with the id from the init container.
But e.g. ldap has the id 100.
The additionalMount initContainer has no permissions for the local config in this case.